+++ src/strategies/base.ts
@@ -41,11 +41,13 @@ import {GenericJson} from '../updaters/generic-json';
 import {GenericXml} from '../updaters/generic-xml';
 import {PomXml} from '../updaters/java/pom-xml';
 import {GenericYaml} from '../updaters/generic-yaml';
+import {GenericToml} from '../updaters/generic-toml';
 
 const DEFAULT_CHANGELOG_PATH = 'CHANGELOG.md';
 
 export interface BuildUpdatesOptions {
   changelogEntry: string;
+  commits?: ConventionalCommit[];
   newVersion: Version;
   versionsMap: VersionsMap;
   latestVersion?: Version;
@@ -313,6 +315,7 @@ export abstract class BaseStrategy implements Strategy {
       newVersion,
       versionsMap,
       latestVersion: latestRelease?.tag.version,
+      commits: conventionalCommits,
     });
     const updatesWithExtras = mergeUpdates(
       updates.concat(...(await this.extraFileUpdates(newVersion, versionsMap)))
@@ -394,6 +397,13 @@ export abstract class BaseStrategy implements Strategy {
                 path: this.addPath(path),
                 createIfMissing: false,
                 updater: new GenericYaml(extraFile.jsonpath, version),
+              });
+              break;
+            case 'toml':
+              extraFileUpdates.push({
+                path: this.addPath(path),
+                createIfMissing: false,
+                updater: new GenericToml(extraFile.jsonpath, version),
               });
               break;
             case 'xml':