+++ src/strategies/node.ts
@@ -14,12 +14,14 @@
 
 import {BaseStrategy, BuildUpdatesOptions} from './base';
 import {Update} from '../update';
+import {ChangelogJson} from '../updaters/changelog-json';
 import {PackageLockJson} from '../updaters/node/package-lock-json';
 import {SamplesPackageJson} from '../updaters/node/samples-package-json';
 import {Changelog} from '../updaters/changelog';
 import {PackageJson} from '../updaters/node/package-json';
 import {GitHubFileContents} from '@google-automations/git-file-utils';
 import {FileNotFoundError, MissingRequiredFileError} from '../errors';
+import {filterCommits} from '../util/filter-commits';
 
 export class Node extends BaseStrategy {
   private pkgJsonContents?: GitHubFileContents;
@@ -67,6 +69,21 @@ export class Node extends BaseStrategy {
         version,
       }),
     });
+
+    // If a machine readable changelog.json exists update it:
+    if (options.commits && packageName) {
+      const commits = filterCommits(options.commits, this.changelogSections);
+      updates.push({
+        path: 'changelog.json',
+        createIfMissing: false,
+        updater: new ChangelogJson({
+          artifactName: packageName,
+          version,
+          commits,
+          language: 'JAVASCRIPT',
+        }),
+      });
+    }
 
     return updates;
   }