+++ src/strategies/python.ts
@@ -15,6 +15,7 @@
 import {BaseStrategy, BuildUpdatesOptions, BaseStrategyOptions} from './base';
 import {Update} from '../update';
 import {Changelog} from '../updaters/changelog';
+import {ChangelogJson} from '../updaters/changelog-json';
 import {Version} from '../version';
 import {SetupCfg} from '../updaters/python/setup-cfg';
 import {SetupPy} from '../updaters/python/setup-py';
@@ -24,6 +25,8 @@ import {
   PyProjectToml,
 } from '../updaters/python/pyproject-toml';
 import {PythonFileWithVersion} from '../updaters/python/python-file-with-version';
+import {FileNotFoundError} from '../errors';
+import {filterCommits} from '../util/filter-commits';
 
 const CHANGELOG_SECTIONS = [
   {type: 'feat', section: 'Features'},
@@ -133,6 +136,22 @@ export class Python extends BaseStrategy {
       });
     });
 
+    // If a machine readable changelog.json exists update it:
+    const artifactName = projectName ?? (await this.getNameFromSetupPy());
+    if (options.commits && artifactName) {
+      const commits = filterCommits(options.commits, this.changelogSections);
+      updates.push({
+        path: 'changelog.json',
+        createIfMissing: false,
+        updater: new ChangelogJson({
+          artifactName,
+          version,
+          commits,
+          language: 'PYTHON',
+        }),
+      });
+    }
+
     return updates;
   }
 
@@ -145,6 +164,35 @@ export class Python extends BaseStrategy {
       return parsePyProject(content.parsedContent);
     } catch (e) {
       return null;
+    }
+  }
+
+  protected async getNameFromSetupPy(): Promise<string | null> {
+    const ARTIFACT_NAME_REGEX = /name *= *['"](?<name>.*)['"](\r|\n|$)/;
+    const setupPyContents = await this.getSetupPyContents();
+    if (setupPyContents) {
+      const match = setupPyContents.match(ARTIFACT_NAME_REGEX);
+      if (match && match?.groups?.name) {
+        return match.groups.name;
+      }
+    }
+    return null;
+  }
+
+  protected async getSetupPyContents(): Promise<string | null> {
+    try {
+      return (
+        await this.github.getFileContentsOnBranch(
+          this.addPath('setup.py'),
+          this.targetBranch
+        )
+      ).parsedContent;
+    } catch (e) {
+      if (e instanceof FileNotFoundError) {
+        return null;
+      } else {
+        throw e;
+      }
     }
   }